pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Manage Jira issues."
authors = ["Matthias Osswald <m@osswald.li>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
python-dotenv = "^0.15.0"
selenium = "^3.141.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
